requirements.txt
fastapi
google-adk
google-generativeai
google-genai
httpx
python-dotenv
requests
streamlit
uvicorn